Change the algorithm for constraining rx1
Now, layers are thickened only where necessary to achieve rx1 < rx1Ma
in a more careful way, then smoothed (on smoothing iterations beyond
the first).  The results seem to significantly reduce the size of the
"teeth" in the bathymetry that were prevalant using the previous method
in the 2D ice-shelf test case.