Cleanup of shortwave routines
This commit removes the writing of restart_forcing_timestamp unless the
new variable shortwave scheme is activated.  It also sets the default
name to follow the convention of the forward model.

It also changes a flag to true for the spatially variable shortwave scheme, when
config_sw_absorption_type='ohlmann00'