Add tools to support regression suites
This commit replaces run_all_testcases.py with a definitions of regression
suites, and a script to manage regression suites.

A regression suite is an arbitrary set of test cases that should be
setup as part of a suite, where the suite is successfull if all tests
pass.

The manage_testcases.py script is used to clean or setup a set of test
cases. Additionally, it writes a summary of the number of processors /
threads / cores needed for the summary to aid in writing batch queue
submission scripts.