Change cori default queue
The previous queue (debug) should probably not be the default queue
because it can't run acme_developer via create_test (creates too many
jobs in the queue). This commit changes the default to the regular
queue.

Also, change BatchMaker to prefer the default queue if that queue
mathes the job specs.

[BFB]