Do not adjust vertical levels on a restart
There is a bit of code that tries to adjust vertical levels to make sure
they add to 1.0.  However, it seems to always get triggered, perhaps due to
roundoff errors.  In any case, if it was applied on a restart, it would
very slightly change the vertical levels from the original run.  This
could lead to non-BFB errors eventually (first found when using the
CP2010 flow A calculation).

It may be better to remove this adjustment entirely and simply replace
it with a check within some tolerance.