Add baseline comparisions
This commit updates the scripts related to setting up test cases to
allow for baseline comparisions. Both the manage_regression_suite.py and
setup_testcase.py scripts now have a -b (--baseline_dir) input argument
that is optional to specify where baselines are stored to compare
against.

When it is provided, any comparision of fields will additionally
generate the same comparision for an identical file in the baseline
location.

For example:

<compare_fields file1="case1/blah.nc" file2="case2/blah.nc">
    <field name="myField" l1_norm="0.0"/>
</compare_fields>

Will typically compare file1 and file2 and ensure that the l1 norm of
the difference for field 'myField' has a value of no more than 0.0.

When baseline comparisions are turned on, file1 will additionally be
compared to file1 in the baseline directory, as will file2 (with file2
instead though).