Remove ALE call in stage 1 of split explicit.
This call was used to compute vertical transport, used for the velocity
tendencies in the baroclinic velocity stage (stage 1).  It turns out
that the ALE call is not very important, and can be substituted with:
 - set vertical transport to zero on first iteration
 - use vertical transport from previous iteration after that.

The advantage is that the ALE routine is now only called in stage 3, so
it reduces computation, and allows us to compute vertical transport (and
the new ALE thickness) based on the variables available in stage 3.