Adds two single column model test cases
This commit adds two versions of the single column CVMIX setup.  The
first is on a planar periodic grid (16 cells).  The second is on a
sphere (no bottom topography or continents).  This case uses the
QU1920km mesh.

This commit is also easily extendable to other tests (e.g. BGC).