Allow baseline comparisons without multiple files
This commit updates the <compare_fields> tag to allow only a single file
to be specified. This means that the fields listed should only be
compared against a baseline, not against another file within the current
test case. This is useful for smoke tests to ensure answers don't change
between model vesions.

Additionally, updating the driver_script documentation to be correct for
validation steps.