Update regression suite scripts to have unbuffered output
This commit updates the generated regression suite scripts to have
unbuffered output so that testcase log files are written in the correct
order.

Previously, field comparisions (and errors associated with them) showed
up in different orders than the steps of a test case. Making the output
be unbuffered fixes this issue and allows the log files to be easier to
read.