Enable simultaneous organic sea spray + aerosol resuspension
In order to simultaneously enable the marine organic sea spray
emissions and the aerosol resuspension mods, a new chemistry
mechanism is required that includes all of the new species.

This commit implements the new mechanism.  It is automatically
selected when CAM is configured with the options for both the 4-mode
marine organics implementation, and the resuspension modifications,
i.e. by passing the following to CAM_CONFIG_OPTS:

-chem trop_mam4_mom -rain_evap_to_coarse_aero

[BFB]
[FCC]
[AG-143]