Add Albany build instructions for new Land Ice solver
This adds compile and linking details for linking to the Albany solver under MPAS-LI. Albany is an unstructured-grid finite element code developed under the PISCEES SciDAC, that itself depends heavily on the Trilinos solver libraries.

Because Albany+Trilinos is C++ code, and is VERY large, we decided at this point to NOT move these codes under ACME, but to pre-install them as external libraries on all the important platforms. (This decision can be revisited.) So, along with this pull request, I installed Albany+Trilinos on Edison, Mira, Titan, Melvin, and skybrigde/redsky. Mustang/Wolf also have Albany installs. This pull request includes an ALBANY_PATH variable in the config_compilers.xml block for each of these machines that points to the install. (Provenance for future Albany releases to ACME still needs to be figured out.)

Linking ACME Fortran to Albany C++ was straightforward on Edison, melvin, skybridge, but was problematic on Titan and Mira. The C++ compilers that go along with the pgi and ibm compilers on those machines are too old to compile Albany. So, we compiled Albany with g++ on Titan and clang++ on Mira. It was possible to link the mixed languages and mixed compilers when the magic libraries were added to the link lines. This is also in config_compilers.xml.

On mira, I needed to remove -qsmp=omp from LDFLAGS for all IBM compilers, since this was not found when linking with clang++. I replaced this flag with -lxlsmp and the code linked. I ran a B-compset and verified that this did not effect performance for non-Albany problems on Mira.

As part of this pull request, a test was added for the MPASLIALB-ONLY compset.

Design Document:
O_21_LI Third Party Support for MPAS-LI within ACME

[CC] New test
[BFB] Existing tests