Fix interpolation at periodic boundary, north pole
The new interpolation code did not work correctly at the periodic
boundary in longitude because of a mod() that should have been
a modulo().  (The two funcitons behave differently for negative
arguments.)

At the northern boundary, the default was previously to extrapolate
but is now to "clamp" to the nearest value as was done in the
earlier interpolation scheme.

Values of T and S in the global_ocean test case now agree to within
machine roundoff between the previous interpolation scheme and the
new subroutines.