Add the ability for templates to contain templates
This commit updates the way templates are applied, to allow templates to
be bundled together into "meta-templates".

Within a template another template can be used, and it will be applied
the same way it would had it just been used in the top level code to
begin with.

NOTE: This ability also allows infinite loops of template application,
so be careful when applying templates.