Rename globalRx1Max; change defaults in 2D shelf
Changed globalMaxRx1Edge to globalRx1Max to be more consistent
with other global reductions (e.g. in globalStats) and because
the global max of rx1 is the same regardless of whether it is
computed from rx1Cell or rx1Edge.

Changed the default cavity thickness in the sub_ice_shelf_2D test
case to be 25 m, because this works better with the new SSH/SSP
initialization.  The default from before (10 m) crashes during
forward runs because the SSP is not well balanced.  Iterative init
can solve this problem.