Fix incorrect merge conflict resolution for blues+pgi env
This commit fixes the incorrect merge resolution in #e45fc33
that modified the pgi build environment added in #d006da9
for blues (pgi 15.0).

Renaming the pgi 15.0 build configuration to "pgi" (default)
and renaming the older (pgi 13*) configuration to "pgi-13"

This incorrect merge resolution is currently only present on
next and not on master. So this change should only be merged
to next (The integrator would take care to resolve conflicts
to preserve the pgi build env when merging to master, so this
change does not have to go to the master).

Note: THIS CHANGE SHOULD NOT BE MERGED TO MASTER