Add optional 'precision' argument to low-level MPAS_io_put_att* routines
This optional argument is added to the 'real0' and 'real1' attribute routines
and allows one to write real-valued attributes at the specified precision.

At present, it is only possible to write single- and double-precision attributes
when MPAS is compiled in double-precision; when MPAS is compiled in single-precision,
only single-precision attributes can be written.