Added some halo updates required for the new advection scheme
The last two commits gave BFB results (compared to the old advection scheme)
for the dome problem when run in serial, but not in parallel.

With this commit, the new advection scheme gives the same results (BFB)
as the old scheme when both are run on 4 processors on a Mac.

Also, the new advection scheme gives the same results (BFB) on 4 processors
as on 1 processor.

The key changes were the addition of some halo updates at the end of
the advection_solver subroutine.

I also fixed a divzero error in an optional energy conservation check.
(This turned up in debug mode.)

I added some debugging diagnostics as I was working toward BFB results;
these can be removed later.

I verified that this commit passes all the current lettuce tests
(5 features, 11 scenarios, 96 tests). The previous two commits do not.
The three commits could be squashed if it is desired to have all commits
pass lettuce.