Removed tracers var_array structure from thermal pool
Previously, the thermal tracers (temperature, enthalpy and waterfrac)
were part of a var_array data structure that was somewhat clumsy.

Now, temperature, enthalpy and waterfrac are 2D fields that are part
of the thermal pool like other thermal-related fields (surfaceTemperature,
basalTemperature, etc.)

Answers are BFB.  All lettuce tests passed.

These changes require a clean and rebuild to update the Registry.

NOTE: streams.landice must be changed. This line should be removed:

    <var_array name="tracers"/>

and replaced by the following:

    <var name="temperature"/>

To run the lettuce tests, I did the following in tests/testing_inputs:
- Unzipped/untarred the version of dome.tar.gz downloaded from oceans11
- Edited dome/streams.landice
- Retarred/rezipped to get a modified version of dome.tar.gz

The gis test does not include tracers in streams.landice, so I did not
have to change it.

At some point, waterfrac should also be part of the streams.
I am leaving it out for now since we are not yet using the enthalpy heat-balance code.