Added logic to speed up the IR code in ice-free regions
The IR code contains a number of loops over nCells and nEdges.
Loop calculations can often be skipped for cells that are ice-free
and for edges that have zero velocity at both vertices.
Previously, many calculations were done in ice-free regions.
To speed up the code, I inserted logic in several place to check for nonzero
ice area or fluxes.

- The new "maskCell" field is set to 1 for cells where ice is present in
  at least one category and is set to 0 elsewhere.
- The new "maskEdge" field is set to 1 for edges where at least one vertex
  has nonzero velocity.
- During construction of linear tracer fields, several subroutines now do
  computations only where maskCell = 1.
- During subroutine integrate_fluxes_over_triangles and in a couple of other
  subroutines, computations are now done only where maskEdge = 1.

The use of maskCell significantly speeds up the construction of linear fields,
by a factor of nearly 3 for a gx3 test problem.

The use of maskEdge, however, gives only modest gains. It turns out that
subroutine integrate_fluxes_over_triangles (the most expensive subroutine
in the code) already contained logic to limit computations in ice-free areas.

These changes are BFB.