Modified the indentation in the IR module
This commit is strictly cosmetic. In the previous commit, I introduced some new logic
that broke the convention of having 3 indented spaces for each new level of the
logical hierarchy.  I went back and changed the indentation to follow the convention.
Many lines of code are touched, but no changes are significant.

This change is BFB.