Resolved case.build conflicts
Introduced several specific acme project ifs to temporarily get this
also working in cesm. A more general approach will need to be implemented
after the initial merge is complete.

Also changed the cesm config_machines.xml to be able to work with the acme
changes. There are still problems with the cesm build for csm_share at this point.