Modified the new advection scheme to support adaptive time steps
Recent testing showed that the new advection scheme did not support an adaptive time step.
If config_adaptive_timestep = true, then dt is set to zero at the start of each time step.
It was not being recomputed until after calling the thermal solver and the advection solver,
which means that neither temperature nor thickness could evolve.

With this commit, the advection calculation is split into two parts:
- The first part, in subroutine prepare_advection, computes layerNormalVelocity (which is
now a field in the Registry) and uses that information to diagnose CFL info and
determine the adaptive time step.
- The second part, in subroutine advection_solver, does the actual advection.

To be on the safe side, I added logic to abort the thermal calculation if called
with dt = 0.

With this change, the thermal and advection solver use the desired adaptive step.
Answers without an adaptive time step are BFB.