disabling irrelevant t_adj calls
The t_adj routine is disabled in threaded regions when threading is
enabled. Commented out the ACME-relevant calls to t_adj in HOMME,
so that timers in MPI-only and MPI/OpenMP runs are identical.
Also commented out some routines that are called from multiple
locations, but for which the call sites are surrounded by timers.