adding option to prefix timer names with detail level
The profile_detail_limit is used to control what performance
timers are measured, an, for example, is used to disable
detailed timing within the initialization. As ACME is
being developed, it is useful to be able to identify
what detail level is associated with each timing event
(and to identify timing events that occur at different
detail levels). This adds a new timing library namelist
parameter, profile_add_detail, that adds the current detail
level as a prefix to the timer name. This is disabled by
default.