Add calcDiffusivity package to apparentDiffusivity
This is a minor detail that must have existed from way back when that
I noticed in testing this branch.  The variable apparentDiffusivity
is only needed when diffusivity is being calculated, which is controlled
by a package combined with namelist options.  This commit simply makes
the apparentDiffusivity variable active only if the calcDiffusivity
package is enabled.