Fix bugs in hetfrz_classnuc,restart bug in CLUBB,PIO2 incompatibility
This PR fixes bugs reported by NCAR:

1. hetfrz_classnuc: With the original value of pdf_n_theta set to 101
the dust activation fraction between -15 and 0 C could be overestimated.
This problem was eliminated by increasing pdf_n_theta to 301. To reduce
the expense of computing the dust activation fraction the integral is
only evaluated where dim_theta is non-zero. This was determined to be
between dim_theta index values of 53 through 113. These loop bounds are
hardcoded in the variables i1 and i2.

2. restart bug in CLUBB:add optional argument to atm_import which is used
during the initialization phase of a restarted run so that cam_in fields
wsx, wsy, shf, and cflx are not overwritten by fields read in from the rs
restart file. These fields are set by the data from the cam.r. restart
file. When nudging to specified dynamics the nudged surface fields
presist in the cam.r. restart so that CLUBB inputs are bit-for-bit the
same on the first step of a restarted run as a continuous run.

3. PIO bug fix: Fixed but in cam_pio_get_var for initial files with multiple
timeslices. (In physics_buffer.F90.in) Output buffers with a singleton mdim
as 2-D variable for PIO2 compatibility

[BFB]

AG-371, AG-392, AG-394