Fix bug that allowed an output-only stream to be unintentionally read
The logic in MPAS_stream_mgr_read( ) that was intended to read all
"input" or "input,output" streams when a streamID was not provided as
an optional argument incorrectly checked whether the stream being
considered was *not* an "input,output" stream.

This commit corrects the faulty logic so that only "input" or "input,output"
streams are read by MPAS_stream_mgr_read( ).