Simplify how calvingThickness updates thickness
To more naturally accommodate the data calving option, I moved the place
where thickness is updated from calvingThickness to outside of the
if-construct for the various calving options.  So now the if-construct
for the various calving options only calculates calvingThickness;
thickness is then updated based on calvingThickness in one location after the
if-construct.