Enable PIO 2 library to be selected in build command with USE_PIO2=true
Rather than requiring users to edit the top-level Makefile to specify
a CPP flag to enable the use of the PIO 2 library, this commit enables
the PIO 2 library to be used by specifying USE_PIO2 in the build command.
This is simpler for users and it is more consistent with how the use
of other libraries is enabled.