Add mpas_test_core_streams.F module for testing stream I/O
The initial purpose of this module is to test that we can read and
write both single- and double-precision streams. This module contains
a single routine that creates a single-precision stream and a double-precision
stream, writes both streams, then reads back both streams to verify that
we get back the same data that were written.

Ideally, one would compile the test core in both single- and double-precision
and run to verify that every combination of executable precision and stream
precision is working.