Add build option PRECISION={single,double}
When building the code, one may now specify PRECISION=single to build with
default single-precision real kind, or PRECISION=double to build with default
double-precision real kind. If PRECISION is not specified, the default is
a double-precision build.

This new options has been implemented by splitting out the Fortran compiler
flags responsible for the auto-promotion of default real kind to double precision,
and selectively adding these promotion flags when building a double-precision
executable.