Default thermal b.c. fields to come from file
Similar to having the default temperature initialization being from
file, this commit sets config_surface_air_temperature_source and
config_basal_heat_flux_source to 'file', so that by default values will
be taken from an input file.  If a user wants to input constant values
instead, this default can be overridden with "constant".

On a related note, I've also added surfaceAirTemperature and
basalHeatFlux ot the input stream so they can actually be input from a
file.  Note that if those fields are not present in the input file, a
warning is written to the log.*.err file, but no error occurs.