Change 'data' config_basal_mass_bal_float to 'file', update descriptions
To be consistent with other options, the 'data'
config_basal_mass_bal_float option is renamed to 'file'.

This commit also updates the description of a number of namelist options
with more information about the various settings.

Also added an 'else' to the logic for config_basal_mass_bal_float in the
code so that a message is printed if an invalid option is selected.
However, it does not actually generate an error or abort the model.