Always calculate global ACFL and DCFL time steps
Previously there was complicated logic for calculating CFL limits:
* only calculate the global ACFL time step if either the adaptive time
stepper is on or extra advection debug info was enabled
* only calculate diffusivity and local and global DCFL time step if the
adaptive time stepper was enabled AND the optional DCFL check was
enabled.  A package was used to only allocate the apparentDiffusivity
variable if needed.

This commit eliminates all of that logic and always calculates the local
and global ACFL and DCFL time steps.  They will only be used if the
adaptive time stepper is enabled (with the DCFL limit still being
optional).  But this change simplifies the code and simplifies the
options for a user.  The other benefit is the CFL information is always
available for output, which is useful for assessing the chosen static
time step after a simulation.  As such, I have added both CFL limited
time steps to the default output stream - as scalars, there is no real
reason not to output them.

I also added timers around the calculations that were originally deemed
to expensive to have on all the time so that we can assess if there is
a performance penalty to this change.  For a short SIA dome test on 4
processors, total cost is ~1% of simulation.  For a HO velocity solver,
the relative cost of these calculations will be an order of magnitude or
two smaller yet.