Only print flat slope diffusivity notice in verbose config
When the apparentDiffusivity is calculated, there is a notice printed to
the error log if flat slopes are encountered saying they are ignored for
the calculation.  Now that this calculation is always performed, this
can result in a lot of marginally useful output.  This commit causes
that notice to only be printed if the option
config_print_thickness_advection_info is set to true.