Add dome restart test
There are a number of ad hoc ways in which a restart could be run
using this system.  I chose to do it like:
* full_run runs for two time steps
* restart_run runs a cold start for one time step and then does a
restart that runs for another time step.  The output of these two steps
is written in the same output.nc file so that output.nc should be a
complete duplication of what was created in the full_run - part was
written in the cold start and part was written in the restart.

One drawback to this approach is the cold start run in the restart_run dir
is redundant with the first part of the run in full_run so there is an
inefficiency there.

In the ocean core they do things a bit differently which avoids this
inefficiency.  There is a shared restart directory, so full_run writes
restart files into that directory, and then restart_run skips the
cold start phase and just does a restart using the restart files in the
"shared" restart directory.  The reason I chose not to follow this
organization is because then there is not an easy way to compare the
output between full_run and restart_run because the two output.nc files
have differing numbers of time levels in them.  The ocean core has
chosen to deal with that by just comparing the final time of the
simulation.  However, I've found that in diagnosing restart problems,
it's often very helpful to see what differences, if any, exist in the
initial time of the restart before time-stepping occurs, and the
approach the ocean takes would not be comparing that time level.

There are other possible ways this could be setup, for example by
copying output.nc from full_run to restart_run and then using
clobber_mode=overwrite to just over write the time levels run during a
restart.  I chose not to do that because if for some reason the run
didn't write output or something, the comparison would see the original,
inadvertently-unmodified output.nc file and think it passed.  There are
other possibilities using, e.g., ncks to take parts of files.  That might
be cleaner but harder to maintain in that there is another step where you
have to make sure you are lining times up correctly.