Updates to CVMix interface
This pull request makes two changes to the CVMix interface

1) Shear based (gradient richardson number) mixing is extended into the
boundary layer.  In other words it is computed everywhere and added to
the KPP boundary layer diagnosed value.  This is done as the Shear
scheme and KPP scheme represent two different scales of processes

2) the contribution of penetrating shortwave to the buoyancy flux is
altered.  Previously, all of the shortwave radiation incident at the
surface contributed to the surface buoyancy flux.  Now a depth of
shortwave absorption is chosen, which is specified in a namelist
variable.