Made a separate subroutine for computing a linear temperature profile
I created a new subroutine, li_init_linear_temperature_in_column, that can be
called either during initialization (if the linear temperature option is chosen)
or during calving-front restoration (to give the restored ice a sensible profile).

The calculation is the same as done previously for initialization, just moved
to a separate subroutine so that it can also be called from the calving module.

Answers are BFB compared to the previous commit.
I verified that the calving restoration test problem still restarts exactly.