Fixed an erroneous error for the case of no advection
Previously, the advection solver was aborting when both config_thickness_advection
and config_tracer_advection were set to 'none'. The reason was a simple logic error;
err was being set to 1 when subroutine li_advection_thickness_tracers was not called.

With this commit, err is set to 1 only if subroutine li_advection_thickness_tracers
is called and returns an error.

This commit is BFB for standard tests.