Always calculate derived velocity variables
There are a number of fields that are derived from velocity
(uReconstructX, surfaceSpeed, etc.) but not part of the actual velocity
solve.  These are part of the li_velocity_solve routine so they were not
being calculated in the initial time of a restart due to logic
preventing velocity from being re-calculated on the initial time of a
restart (which was necessary for exact restarts).  However, this
resulted in allowableDtDCFL being calculated incorrectly on the first time step
of a restart.

This commit moves the logic of whether to solve velocity or not into the
li_velocity_solve routine, instead of around it.  The logic now only
controls the solution of the velocity field itself, and the derived
velocity fields are always calculated.