Removed unneeded time levels in thickness and layerThickness fields
In the previous commits, I was assuming two time levels for the thickness field,
which in fact has just one time level.  I removed the references to the second time level.

Also, the layerThickness field no longer needs two time levels. I modified the advection
logic so that only one time level is used.

While doing so, I removed a scratch field that was serving as a second time level for tracers.
Thus the layer thickness and tracers are both treated as
   field = field + tendency*dt    (one time level)
rather than
   newField = oldField + tendency*dt     (two time levels)

I also removed 'thickness' from the argument list for the main advection subroutine, since it
is already included in the geometryPool.

Note: The only 2-time-level fields remaining in the Registry are vertexMask and dirichletVelocityMask.

I verified that answers for the dome test are BFB compared to the previous commit,
for thickness-only advection and also for thickness-plus-tracer advection.
(But recall that the previous commit is not BFB compared to landice/develop.)