Fix incorrect timelevels
A recent branch reduced the number of time levels for thickness and
layerThickness from 2 to 1.  This commit fixes a couple instances that
were still referencing the second time level and were not caught in the
previous branch.  In the instance in mpas_li_setup.F, the reference to
layerThickness is not actually needed, so I just removed it.

This eliminates a number of harmless but confusing 'error' messages.