Improved the treatment of tracers at the ice surface
As part of the advection scheme, ice is added or removed at the upper and lower surfaces
through the surface and basal mass balance terms.  Previously, when ice was added,
tracer values in the top and bottom ice layers were not modified.  This means
that the incoming ice was assumed to have the same temperature as the existing ice.
Often this is not a bad approximation, but it is a poor approximation
when the surface air temperature is well below freezing and the SMB is being
added to a previously ice-free cell with temperatures set to 0 C (the default).
For EISMINT-2 simulations with a long timestep (~100 years), the ice was being
warmed excessively by an SMB assumed to arrive with a temperature of 0 C.

With this commit, new surfaceTracer and basalTracer fields are computed in the
advection scheme as part of the tracer setup.  These tracer values (specifically,
temperature or enthalpy) are applied to new ice at the upper and/or lower surface.
The upper- and lower-layer tracer values are then adjusted conservatively,
giving appropriate weights to the existing and new ice.  These calculations
are done in a new subroutine, apply_mass_balance.

Also, I fixed a bug whereby the advection scheme was ignoring tracer advection even when
config_tracer_advection = 'fo', because the logical variable advectTracers was set to
false by default.  Not sure how this slipped through previous commits.

I added an option to config_temperature_init: The initial temperature can now
be set uniformly to the surface air temperature in addition to being read from a file or
set to a linear profile.

I added a couple of scratch fields to the registry to handle the temporary
surfaceTracer and basalTracer fields.

This commit should be BFB for standard tests with thickness-only advection, but will
change the answers when tracer advection is turned on.