Update dome setup script
Recently the create LI grid script had temperature removed as a default
field to be created in a LI mesh.  This change eliminates temperature
from the dome script since it causes an error to look for it - and it's
not needed for this test anyway.

Additionally, I did some other minor cleanup.