changing inheriting of _base modules in HOMME
It was requested that everything public in the _base modules be explicitly listed in the inheriting modules of each target. So I have changed preqx to list them all as a template for new solvers to use. I've also made the new modules private so they need to re-declare public everything in the base module and the new solver module inheriting it.