Ensure Okubo Weiss timers exist on all procs
This commit updates the Okubo Weiss analysis member to ensure all timers
are started / stopped on all processors. If they are not, the code
hangs when trying to synchronize all of the timers at the end of the
run.

This issue can be seen in the 20km ZISO test case using it's default
configuration.