Set path to new SOAG emission file as soag_ext_file
Set path to new SOAG emission file as soag_ext_file instead of as
soag_emis_file, because it is an elevated, as opposed to a surface,
emission dataset.  This requires modifications to CAM's build-namelist
that associate the emission file path to the correct tracer name.

This commit changes namelists in the new compsets (which are not
currently in the testing suite).

[BFB]
[FCC]

See confluence for a more detailed description about these tags.