Adds a surface buoyancy computation fix
This commit adds a few changes in how the surface buoyancy flux is
computed for input into KPP.  Surface thickness inputs are now added
into layer 1.  A few additions to the surface tracer flux are added.  1)
previously, when thickness was decreased in the model, the temperature
would change according to the latent heat flux and the thickness change.
Thus the temperature change was too strong.  In this commit the
temperature is changed via latent heat flux only.  2) river runoff is
added to the ocean at the local temperature of the ocean, instead of
zero degrees