Fix MPI comm 'leak' in Interface
The createReducedMPI function was creating a new communicator with each
call (i.e. one per time step).  This eventually causes problems as many
machines have a limit to the number of communicators allowed.

This fix checks if the communicator already exists and if so it
frees it before creating a new one.