Fix linking of executable, and add run model messages
This commit updates linking of the executable in a <model_run> step to
cause the run scripts to actually run the executable in the given
directory, rather than using a full link back to the original
executable.

Additionally, this commit adds printing of start/finish messages for
model run steps, to more clearly define when a model run is occurring.