Change coupling sequencer and time steps
This commit updates the default configuration for the low resolution
water cycle simulation to help resolve the intertial instability issues
related to the ocean/sea-ice models.

Notable changes are that the coupling sequencer is now RASM_OPTION1, and
the default coupling interval for the atmosphere/sea-ice/ocean models
are set to 30 minutes.

[non-BFB]