Change computation of Bulk Richardson Number
This commit makes a very small change in the computation of the bulk
Richardson number used to calculate the boundary layer depth in KPP.
Previously, the in-situ density was used to determine the density
difference between a large, boundary layer eddy and the local density.
As these eddies may traverse a large distance, it makes sense to use
the potentialDensity instead, as this is conserved for adiabatic
displacements.