Fixes a bug in the addition of background diffusivity/viscosity
In the CVMix interface, the background viscosity/diffusivity is added
to the input viscosity/diffusivity array in CVMix, after returning from
CVMix, the background diffusivity is added again for SimpleShapes, but not
MatchBoth.  If the CVMix code is examined, the input arrays are added to what
is computed within the calls to CVMix.  Therefore, the addition of
background viscosity/diffusivity a second time doubles the value entered in the
namelist file.

This commit removes the second addition of background diffusivity/viscosity